+++ src/simmate/apps/chemspace/components/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from .chem_space_freedom_space_molecule_table import ChemSpaceFreedomSpaceMoleculeTable
+from .freedom_space_molecule import ChemSpaceFreedomSpaceMoleculeComponent