+++ src/simmate/apps/eppo_gd/components/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from .eppo_code_table import EppoCodeTable
+from .code import EppoCodeComponent