+++ src/simmate/apps/pdb/components/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from .pdb_ligand_table import PdbLigandTable
+from .ligand import PdbLigandComponent