+++ src/simmate/compute/components/__init__.py
@@ -1,4 +1,4 @@
 # -*- coding: utf-8 -*-
 
-from .simmate_worker_table import SimmateWorkerTable
-from .work_item_table import WorkItemTable
+from .simmate_worker import SimmateWorkerComponent
+from .work_item import WorkItemComponent