+++ src/simmate/website/data_explorer/components/__init__.py
@@ -1,6 +1,3 @@
 # -*- coding: utf-8 -*-
 
 from .table_component import TableComponent
-
-# DEPRECATED: use TableComponent instead
-DynamicTableForm = TableComponent