+++ src/simmate/website/data_explorer/test/test_data_explorer_views.py
@@ -50,4 +50,4 @@ def test_workflows_by_type_view(client, table_name):
 
     response = client.get(url)
     assert response.status_code == 404
-    # assertTemplateUsed(response, "data_explorer/table_entry.html")
+    # assertTemplateUsed(response, "data_explorer/entry.html")